requirements.txt
fastapi>=0.110,<1.0
uvicorn[standard]>=0.29,<1.0
sqlalchemy>=2.0,<3.0
jinja2>=3.1,<4.0
PyYAML>=6.0,<7.0
packaging>=24.0,<25.0

